Implement adding meal to diary and remove from daily calories

+++ Data/StayFit.Data.Models/MealDiary.cs
@@ -12,6 +12,6 @@
 
         public virtual ApplicationUser User { get; set; }
 
-        public decimal MealQuantity { get; set; }
+        public double MealQuantity { get; set; }
     }
 }
